requirements.txt
streamlit==1.28.0
pdfplumber==0.10.3
pandas==2.1.0
python-dateutil==2.8.2